test dis_001

+++ dis_001_significant_earthquakes/contents/src/__init__.py
@@ -2,45 +2,50 @@ import logging
 import sys
 import os
 import time
-import urllib.request
+import requests as req
 from collections import OrderedDict
 from datetime import datetime, timedelta
 import cartosql
 
 ### Constants
-SOURCE_URL = "ftp://podaac.jpl.nasa.gov/allData/merged_alt/L2/TP_J1_OSTM/global_mean_sea_level/"
-FILENAME_INDEX = -1
-DATETIME_INDEX = 2
-TIMEOUT = 300
-ENCODING = 'utf-8'
-STRICT = False
+SOURCE_URL = "https://earthquake.usgs.gov/fdsnws/event/1/query?format=geojson&starttime={startTime}&endtime={endTime}&minsig={minSig}"
+
+PROCESS_HISTORY = True
+DATE_FORMAT = '%Y-%m-%d'
+DATETIME_FORMAT = '%Y-%m-%dT%H:%M:%SZ'
+SIGNIFICANT_THRESHOLD=600
+
+LOG_LEVEL = logging.DEBUG
 CLEAR_TABLE_FIRST = False
 
 ### Table name and structure
-CARTO_TABLE = 'cli_040_mean_sea_level_rise'
+CARTO_TABLE = 'dis_001_significant_earthquakes'
 CARTO_SCHEMA = OrderedDict([
-    ('altimeter_type', 'numeric'),
-    ('merged_file_cycle', 'numeric'),
-    ('date', 'timestamp'),
-    ('num_obs', 'numeric'),
-    ('num_weighted_obs', 'numeric'),
-    ('gmsl_no_gia', 'numeric'),
-    ('sd_gmsl_no_gia', 'numeric'),
-    ('gauss_filt_gmsl_no_gia', 'numeric'),
-    ('gmsl_gia', 'numeric'),
-    ('sd_gmsl_gia', 'numeric'),
-    ('gauss_filt_gmsl_gia', 'numeric'),
-    ('gauss_filt_gmsl_gia_ann_signal_removed', 'numeric')
+    ('uid', 'text'),
+    ('the_geom', 'geometry'),
+    ('depth_in_km', 'numeric'),
+    ('datetime', 'timestamp'),
+    ('mag', 'numeric'),
+    ('place', 'text'),
+    ('sig', 'numeric'),
+    ('magType', 'text'),
+    ('nst', 'numeric'),
+    ('dmin', 'numeric'),
+    ('rms', 'numeric'),
+    ('gap', 'numeric'),
+    ('tsunami', 'numeric'),
+    ('felt', 'numeric'),
+    ('cdi', 'numeric'),
+    ('mmi', 'numeric'),
+    ('net', 'text'),
+    ('alert', 'text')
 ])
-UID_FIELD = 'date'
-TIME_FIELD = 'date'
-
-CARTO_USER = os.environ.get('CARTO_USER')
-CARTO_KEY = os.environ.get('CARTO_KEY')
+UID_FIELD = 'uid'
+TIME_FIELD = 'datetime'
 
 # Table limits
 MAX_ROWS = 1000000
-MAX_AGE = datetime.today() - timedelta(days=365*150)
+MAX_AGE = datetime.today() - timedelta(days=365*5)
 
 ###
 ## Carto code
@@ -60,145 +65,104 @@ def checkCreateTable(table, schema, id_field, time_field):
         if id_field != time_field:
             cartosql.createIndex(table, time_field)
 
-def cleanOldRows(table, time_field, max_age, date_format='%Y-%m-%d %H:%M:%S'):
-    '''
-    Delete excess rows by age
-    Max_Age should be a datetime object or string
-    Return number of dropped rows
-    '''
-    num_expired = 0
-    if cartosql.tableExists(table):
-        if isinstance(max_age, datetime):
-            max_age = max_age.strftime(date_format)
-        elif isinstance(max_age, str):
-            logging.error('Max age must be expressed as a datetime.datetime object')
+def deleteExcessRows(table, max_rows, time_field, max_age=''):
+    '''Delete excess rows by age or count'''
+    num_dropped = 0
+    if isinstance(max_age, datetime):
+        max_age = max_age.isoformat()
 
+    # 1. delete by age
+    if max_age:
         r = cartosql.deleteRows(table, "{} < '{}'".format(time_field, max_age))
-        num_expired = r.json()['total_rows']
-    else:
-        logging.error("{} table does not exist yet".format(table))
+        num_dropped = r.json()['total_rows']
 
-    return(num_expired)
-
-def deleteExcessRows(table, max_rows, time_field):
-    '''Delete rows to bring count down to max_rows'''
-    num_dropped=0
-    # 1. get sorted ids (old->new)
-    r = cartosql.getFields('cartodb_id', table, order='{} desc'.format(time_field),
+    # 2. get sorted ids (old->new)
+    r = cartosql.getFields('cartodb_id', table, order='{}'.format(time_field),
                            f='csv')
     ids = r.text.split('\r\n')[1:-1]
 
-    # 2. delete excess
+    # 3. delete excess
     if len(ids) > max_rows:
-        r = cartosql.deleteRowsByIDs(table, ids[max_rows:])
+        r = cartosql.deleteRowsByIDs(table, ids[:-max_rows])
         num_dropped += r.json()['total_rows']
     if num_dropped:
         logging.info('Dropped {} old rows from {}'.format(num_dropped, table))
-
-    return(num_dropped)
+    return num_dropped
 
 ###
 ## Accessing remote data
 ###
 
-def fetchDataFileName(SOURCE_URL):
-    """
-    Select the appropriate file from FTP to download data from
-    """
-    with urllib.request.urlopen(SOURCE_URL) as f:
-        ftp_contents = f.read().decode('utf-8').splitlines()
+def genUID(lat,lon,depth,dt):
+    return '{}_{}_{}_{}'.format(lat,lon,depth,dt)
 
-    filename = ""
-    ALREADY_FOUND=False
-    for fileline in ftp_contents:
-        fileline = fileline.split()
-        potential_filename = fileline[FILENAME_INDEX]
-        if (potential_filename.endswith(".txt") and ("V4" in potential_filename)):
-            if not ALREADY_FOUND:
-                filename = potential_filename
-                ALREADY_FOUND=True
-            else:
-                logging.warning("There are multiple filenames which match criteria, passing most recent")
-                filename = potential_filename
+def appendTimeFrame(existing_ids, startTime, endTime, new_data, new_ids):
+    query = SOURCE_URL.format(startTime=startTime, endTime=endTime, minSig=SIGNIFICANT_THRESHOLD)
+    res = req.get(query).json()
+    num_start_ids = len(new_ids)
+    for feature in res['features']:
+        coords = feature['geometry']['coordinates']
+        lat = coords[1]
+        lon = coords[0]
+        depth = coords[2]
+        geom = {
+            'type':'Point',
+            'coordinates':[lon,lat]
+        }
 
-    logging.info("Selected filename: {}".format(filename))
-    if not ALREADY_FOUND:
-        logging.warning("No valid filename found")
+        props = feature['properties']
+        dt = datetime.utcfromtimestamp(props['time']/1000).strftime(DATETIME_FORMAT)
 
-    return(filename)
+        _uid = genUID(lat,lon,depth,dt)
+        if _uid not in existing_ids + new_ids:
+            new_ids.append(_uid)
+            row = []
+            for field in CARTO_SCHEMA:
+                if field == 'uid':
+                    row.append(_uid)
+                elif field == 'the_geom':
+                    row.append(geom)
+                elif field == 'depth_in_km':
+                    row.append(depth)
+                elif field == 'datetime':
+                    row.append(dt)
+                else:
+                    row.append(props[field])
 
-def tryRetrieveData(SOURCE_URL, filename, TIMEOUT, ENCODING):
-    # Optional logic in case this request fails with "unable to decode" response
-    start = time.time()
-    elapsed = 0
-    resource_location = os.path.join(SOURCE_URL, filename)
+            new_data.append(row)
 
-    while elapsed < TIMEOUT:
-        elapsed = time.time() - start
-        try:
-            with urllib.request.urlopen(resource_location) as f:
-                res_rows = f.read().decode(ENCODING).splitlines()
-                return(res_rows)
-        except:
-            logging.error("Unable to retrieve resource on this attempt.")
-            time.sleep(5)
+    logging.debug('{} new ids added'.format(len(new_ids) - num_start_ids))
+    return new_data, new_ids
 
-    logging.error("Unable to retrive resource before timeout of {} seconds".format(TIMEOUT))
-    if STRICT:
-        raise Exception("Unable to retrieve data from {}".format(resource_locations))
-    return([])
-
-# https://stackoverflow.com/questions/20911015/decimal-years-to-datetime-in-python
-def decimalToDatetime(dec, date_pattern="%Y-%m-%d %H:%M:%S"):
-    """
-    Convert a decimal representation of a year to a desired string representation
-    I.e. 2016.5 -> 2016-06-01 00:00:00
-    """
-    year = int(dec)
-    rem = dec - year
-    base = datetime(year, 1, 1)
-    dt = base + timedelta(seconds=(base.replace(year=base.year + 1) - base).total_seconds() * rem)
-    result = dt.strftime(date_pattern)
-    return(result)
-
-def insertIfNew(newUID, newValues, existing_ids, new_data):
-    '''
-    For new UID, values, check whether this is already in our table
-    If not, add it
-    Return new_ids and new_data
-    '''
-    seen_ids = existing_ids + list(new_data.keys())
-    if newUID not in seen_ids:
-        new_data[newUID] = newValues
-        logging.debug("Adding {} data to table".format(newUID))
-    else:
-        logging.debug("{} data already in table".format(newUID))
-    return(new_data)
-
-def processData(SOURCE_URL, filename, existing_ids):
+def processData(existing_ids):
     """
     Inputs: FTP SOURCE_URL and filename where data is stored, existing_ids not to duplicate
     Actions: Retrives data, dedupes and formats it, and adds to Carto table
     Output: Number of new rows added
     """
-    num_new = 0
+    new_data = []
+    new_ids = []
 
-    res_rows = tryRetrieveData(SOURCE_URL, filename, TIMEOUT, ENCODING)
-    new_data = {}
-    for row in res_rows:
-        if not (row.startswith("HDR")):
-            row = row.split()
-            if len(row)==len(CARTO_SCHEMA):
-                logging.debug("Processing row: {}".format(row))
-                date = decimalToDatetime(float(row[DATETIME_INDEX]))
-                row[DATETIME_INDEX] = date
-                new_data = insertIfNew(date, row, existing_ids, new_data)
-            else:
-                logging.debug("Skipping row: {}".format(row))
+    today = datetime.today()
 
-    if len(new_data):
-        num_new += len(new_data)
-        new_data = list(new_data.values())
+    if PROCESS_HISTORY:
+
+        startTime = MAX_AGE
+        endTime = startTime + timedelta(days=31)
+        while startTime < today:
+            logging.debug('Fetching data between {} and {}'.format(startTime, endTime))
+            new_data, new_ids = appendTimeFrame(existing_ids, startTime, endTime, new_data, new_ids)
+            startTime = endTime
+            endTime = startTime + timedelta(days=31)
+
+    else:
+        # Use defaults of endpoint
+        startTime = ''
+        endTime = ''
+        new_data = appendTimeFrame(existing_ids, startTime, endTime, new_data, new_ids)
+
+    num_new = len(new_data)
+    if num_new:
         cartosql.blockInsertRows(CARTO_TABLE, CARTO_SCHEMA.keys(), CARTO_SCHEMA.values(), new_data)
 
     return(num_new)
@@ -208,31 +172,27 @@ def processData(SOURCE_URL, filename, existing_ids):
 ###
 
 def main():
-    logging.basicConfig(stream=sys.stderr, level=logging.INFO)
+    logging.basicConfig(stream=sys.stderr, level=LOG_LEVEL)
 
     if CLEAR_TABLE_FIRST:
-        cartosql.dropTable(CARTO_TABLE)
+        if cartosql.tableExists(CARTO_TABLE):
+            cartosql.dropTable(CARTO_TABLE)
 
     ### 1. Check if table exists, if not, create it
     checkCreateTable(CARTO_TABLE, CARTO_SCHEMA, UID_FIELD, TIME_FIELD)
 
-    ### 2. Delete old rows
-    num_expired = cleanOldRows(CARTO_TABLE, TIME_FIELD, MAX_AGE)
-
-    ### 3. Retrieve existing data
+    ### 2. Retrieve existing data
     r = cartosql.getFields(UID_FIELD, CARTO_TABLE, order='{} desc'.format(TIME_FIELD), f='csv')
     existing_ids = r.text.split('\r\n')[1:-1]
     num_existing = len(existing_ids)
 
-    logging.debug("First 10 IDs already in table: {}".format(existing_ids[:10]))
+    ### 3. Fetch data from FTP, dedupe, process
+    num_new = processData(existing_ids)
 
-    ### 4. Fetch data from FTP, dedupe, process
-    filename = fetchDataFileName(SOURCE_URL)
-    num_new = processData(SOURCE_URL, filename, existing_ids)
+    ### 4. Delete data to get back to MAX_ROWS
+    num_dropped = deleteExcessRows(CARTO_TABLE, MAX_ROWS, TIME_FIELD, MAX_AGE)
 
-    ### 5. Delete data to get back to MAX_ROWS
-    num_deleted = deleteExcessRows(CARTO_TABLE, MAX_ROWS, TIME_FIELD)
-
-    ### 6. Notify results
-    logging.info('Expired rows: {}, Previous rows: {},  New rows: {}, Dropped rows: {}, Max: {}'.format(num_expired, num_existing, num_new, num_deleted, MAX_ROWS))
+    ### 5. Notify results
+    total = num_existing + num_new - num_dropped
+    logging.info('Existing rows: {},  New rows: {}, Max: {}'.format(total, num_new, MAX_ROWS))
     logging.info("SUCCESS")
